Reshaped columns in Planning::lowerListView(). Removed botched Logitem.toDollarString() in favor of correct Datademunger.toCurrency(). Planted comment  in  "case 1" forks on Planning..export sections: need dialog to ask if General Allocations should be added (simple yes/no "AlertDialog"). May need to ask similar question on imports too.Marker "*GENERAL ALLOCATION*" proposed
Forgot that askCupertinoDate() returns ISO dates, and added a check as to whether or not to convert getDateButton()'s output.
Can't seem to let this one go for the day...armed some diagnostics. Changed columns on lower Planning page.

Got it right on DB and faux toolbar, trying to cause change of state on LoggingItem edit page

Whoops...mis-IDd a DB field on import operation, and missed completely on setState() for iOS datebuttons

First draft of Planning import/export -- does not include gross allocations, call it a final PoC.
Attempt at balancing the faux toolbar on iOS
Fixed a mistake in gigUI that resulted in Pages not being queried for popupItems.
Corrected gigUI to hopefully ditch the cupertinoAdder when the page warrants. All that remains now is to import/export the Planning data

Forgot to redo LoggingPage's date range button handlers for a coherent state situation

reworking of iOS export

Reorganized Logging Page into separate file, gigUI() in main now asks the current page about add button and (in Android) the Popup Menu. Pages themselves answer for the faux iOS Toolbar with CupertinoTabBar

iOS fixes:
-Logging page now same for both platforms
-Attempt to cause getting of Stats when that page is active
-Aim cupertinoAdder to call the current Page's fabClicked

Made toDollarString() return a string of the value actually rounded to the nearest penny, thus stopping a "fraction from hell" display

Need to handle "unknown category" with a SimpleDialog.